// frontend/src/app/core/services/api/model/scoringSchemeRequest.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ScoringType } from './scoringType';


/**
 * Request payload for creating or updating a scoring scheme
 */
export interface ScoringSchemeRequest { 
    name: string;
    scoringType: ScoringType;
    isActive: boolean;
}
export namespace ScoringSchemeRequest {
}