// frontend/src/app/core/services/api/model/scoringType.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Type of scoring calculation
 */
export const ScoringType = {
    Fixed: 'FIXED',
    OffsetFromEnd: 'OFFSET_FROM_END'
} as const;
export type ScoringType = typeof ScoringType[keyof typeof ScoringType];